requirements.txt
Flask
requests
flask-cors
python-dotenv
gunicorn==21.2.0